changed gpu tests from uint64_t to uint32_t

+++ gpu/test/build-bst-cycles.cu
@@ -22,8 +22,8 @@ int main(int argc, char *argv[]) {
         exit(1);
     }
 
-    uint32_t d = atoi(argv[1]);
-    uint64_t n = pow(2, d) - 1;
+    int d = atoi(argv[1]);
+    int n = pow(2, d) - 1;
     #ifdef DEBUG
     printf("n = 2^%d - 1 = %lu\n", d, n);
     #endif
@@ -31,18 +31,18 @@ int main(int argc, char *argv[]) {
     int p = atoi(argv[2]);
     omp_set_num_threads(p);
 
-    uint64_t *A = (uint64_t *)malloc(n * sizeof(uint64_t));
-    uint64_t *dev_A;
-    cudaMalloc(&dev_A, n * sizeof(uint64_t));
+    int *A = (int *)malloc(n * sizeof(int));
+    int *dev_A;
+    cudaMalloc(&dev_A, n * sizeof(int));
 
     double time[ITERS];
 
     //Construction
-    for (uint32_t i = 0; i < ITERS; ++i) {
-        initSortedList<uint64_t>(A, n);
-        cudaMemcpy(dev_A, A, n * sizeof(uint64_t), cudaMemcpyHostToDevice);
-        time[i] = timePermuteBtree<uint64_t>(dev_A, n, 1);
-        cudaMemcpy(A, dev_A, n * sizeof(uint64_t), cudaMemcpyDeviceToHost);
+    for (int i = 0; i < ITERS; ++i) {
+        initSortedList<int>(A, n);
+        cudaMemcpy(dev_A, A, n * sizeof(int), cudaMemcpyHostToDevice);
+        time[i] = timePermuteBtree<int>(dev_A, n, 1);
+        cudaMemcpy(A, dev_A, n * sizeof(int), cudaMemcpyDeviceToHost);
     }
     printTimings(n, time);
