+++ gpu/test/query-btree-fixedN.cu
@@ -23,32 +23,32 @@ int main(int argc, char *argv[]){
         exit(1);
     }
 
-    uint64_t n = atol(argv[1]);
-    uint64_t b = atoi(argv[2]);
+    int n = atoi(argv[1]);
+    int b = atoi(argv[2]);
     #ifdef DEBUG
-    printf("n = %lu; b = %lu\n", n, b);
+    printf("n = %d; b = %d\n", n, b);
     #endif
 
-    uint64_t p = atoi(argv[3]);
+    int p = atoi(argv[3]);
     omp_set_num_threads(p);
 
-    uint64_t *A = (uint64_t *)malloc(n * sizeof(uint64_t));
-    uint64_t *dev_A;
-    cudaMalloc(&dev_A, n * sizeof(uint64_t));
+    int *A = (int *)malloc(n * sizeof(int));
+    int *dev_A;
+    cudaMalloc(&dev_A, n * sizeof(int));
 
     double time[ITERS];
 
     //Construction
-    initSortedList<uint64_t>(A, n);
-    cudaMemcpy(dev_A, A, n * sizeof(uint64_t), cudaMemcpyHostToDevice);
-    timePermuteBtree<uint64_t>(dev_A, n, b);
-    cudaMemcpy(A, dev_A, n * sizeof(uint64_t), cudaMemcpyDeviceToHost);
+    initSortedList<int>(A, n);
+    cudaMemcpy(dev_A, A, n * sizeof(int), cudaMemcpyHostToDevice);
+    timePermuteBtree<int>(dev_A, n, b);
+    cudaMemcpy(A, dev_A, n * sizeof(int), cudaMemcpyDeviceToHost);
 
     //Querying
-    uint64_t q = 1000000;
+    int q = 1000000;
     while (q <= 100000000) {
-        for (uint32_t i = 0; i < ITERS; ++i) {
-            time[i] = timeQueryBtree<uint64_t>(A, dev_A, n, b, q);
+        for (int i = 0; i < ITERS; ++i) {
+            time[i] = timeQueryBtree<int>(A, dev_A, n, b, q);
         }
         printQueryTimings(n, q, time);
